+++ src/tt.c
@@ -17,7 +17,7 @@ void tt_store(uint64_t hash, int score, int depth, int flag, Move *best) {
 	TTEntry *e = &tt[hash % TT_SIZE];
 	if (e->key == 0 || e->depth <= depth) {
 		e->key = hash;
-		e->score = (int16_t)score;
+		e->score = (int32_t)score;
 		e->depth = (int8_t)depth;
 		e->flag = flag;
 		if (best) e->best = *best;
